Add files via upload
Write an application that reads the radius and height of a cone then calculates and displays the volume and surface area. Use the following formulas, in which r and h represent the cone’s radius and height. Print the output to four decimal places.
Volume = 1/3 π r2 h  Surface Area = π r (r + s) where s is the ‘slant height’ which is √r2+h2