Add files via upload
Design and implement an application that repeatedly reads some text from the user, then determines and displays how many of each of the even digits (i.e., 0, 2, 4, 6, and 8) appear in the entire text. Have a separate counter for each digit. Also count and print the number of all other characters. The logic must include a switch-case structure to determine which counter to increment. After the results for an input are displayed, prompt the user if they want to do another input.

Hint: Use the charAt method of the String class to examine each character of the input.